refactor: restructure the project
