--- src/graphback-core/config.json
@@ -1,3 +1,0 @@
-{
-  "database": "./src/"
-}
